fix(clippy): loc.id() is already i64 -- drop the no-op cast
CI's only red. The tracker's neighbouring reads cast to u64, which is a real
conversion; this one was copied from there and casts i64 to i64.

+++ crates/eldenring-archipelago/src/core.rs
@@ -3480,7 +3480,7 @@ impl Core {
         if let Some(client) = self.client() {
             connected = true;
             for loc in client.checked_locations() {
-                checked.insert(loc.id() as i64);
+                checked.insert(loc.id());
             }
             for _ in client.unchecked_locations() {
                 unchecked_n += 1;
